requirements.txt
# PDF Extraction
camelot-py[cv]

# Data Processing
pandas>=1.5.0
numpy>=1.21.0
openpyxl>=3.0.10
PyPDF2>=3.0.0
tabula-py>=2.5.0

# Machine Learning
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=1.4.0

# Web Framework
streamlit>=1.15.0
plotly>=5.11.0
folium>=0.12.0

# Utilities
requests>=2.28.0
beautifulsoup4>=4.11.0
python-dotenv>=0.19.0

# Development
jupyter>=1.0.0
pytest>=7.0.0
black>=22.0.0
